# board-service/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.BoardResponse:
    properties:
      assignee:
        $ref: '#/definitions/dto.UserInfo'
      author:
        $ref: '#/definitions/dto.UserInfo'
      board_id:
        type: string
      content:
        type: string
      createdAt:
        type: string
      dueDate:
        type: string
      importance:
        $ref: '#/definitions/dto.CustomImportanceResponse'
      project_id:
        type: string
      roles:
        items:
          $ref: '#/definitions/dto.CustomRoleResponse'
        type: array
      stage:
        $ref: '#/definitions/dto.CustomStageResponse'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CreateBoardRequest:
    properties:
      assignee_id:
        type: string
      content:
        maxLength: 5000
        type: string
      dueDate:
        description: ISO 8601 format
        type: string
      importance_id:
        type: string
      project_id:
        type: string
      role_ids:
        items:
          type: string
        minItems: 1
        type: array
      stage_id:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - project_id
    - role_ids
    - stage_id
    - title
    type: object
...
  dto.CustomImportanceResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      displayOrder:
        type: integer
      importance_id:
        type: string
      isSystemDefault:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CustomRoleResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      displayOrder:
        type: integer
      isSystemDefault:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      updatedAt:
        type: string
    type: object
  dto.CustomStageResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      displayOrder:
        type: integer
      isSystemDefault:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      stage_id:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          message:
            type: string
        type: object
      request_id:
        type: string
    type: object
...
  dto.PaginatedBoardsResponse:
    properties:
      boards:
        items:
          $ref: '#/definitions/dto.BoardResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
      request_id:
        type: string
    type: object
  dto.UpdateBoardRequest:
    properties:
      assignee_id:
        type: string
      content:
        maxLength: 5000
        type: string
      dueDate:
        type: string
      importance_id:
        type: string
      role_ids:
        items:
          type: string
        type: array
      stage_id:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
  dto.UserInfo:
    properties:
      email:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@wealist.com
    name: API Support
  description: Board management API for weAlist project management platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Board Service API
  version: "1.0"
paths:
  /api/boards:
    get:
      consumes:
      - application/json
      description: Get boards for a project with optional filters
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      - description: Filter by Stage ID
        in: query
        name: stage_id
        type: string
      - description: Filter by Role ID
        in: query
        name: role_id
        type: string
      - description: Filter by Importance ID
        in: query
        name: importance_id
        type: string
      - description: Filter by Assignee ID
        in: query
        name: assignee_id
        type: string
      - description: Filter by Author ID
        in: query
        name: author_id
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedBoardsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get boards
      tags:
      - boards
    post:
      consumes:
      - application/json
      description: Create a new board card (task/issue) in a project
      parameters:
      - description: Board details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BoardResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create board
      tags:
      - boards
  /api/boards/{board_id}:
    delete:
      consumes:
      - application/json
      description: Delete a board (soft delete, author or ADMIN+ only)
      parameters:
      - description: Board ID
        in: path
        name: board_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete board
      tags:
      - boards
    get:
      consumes:
      - application/json
      description: Get a specific board by ID (project member only)
      parameters:
      - description: Board ID
        in: path
        name: board_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BoardResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get board
      tags:
      - boards
    put:
      consumes:
      - application/json
      description: Update a board (author or ADMIN+ only)
      parameters:
      - description: Board ID
        in: path
        name: board_id
        required: true
        type: string
      - description: Board updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.BoardResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update board
      tags:
      - boards
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"